+++ include/egt/combo.h
@@ -193,7 +193,7 @@ public:
 
 protected:
 
-    void set_parent(Frame* parent) override;
+    void set_parent(Widget* parent) override;
 
     /// Internal list.
     ListBox m_list;
