+++ include/egt/label.h
@@ -114,7 +114,7 @@ public:
 
 protected:
 
-    void set_parent(Frame* parent) override;
+    void set_parent(Widget* parent) override;
 };
 
 /**
