+++ include/egt/radial.h
@@ -169,6 +169,7 @@ public:
             damage();
         }
     }
+    using Widget::remove;
 
     using Widget::color;
