+++ include/egt/widget.h
@@ -11,6 +11,7 @@
  * @brief Base class Widget definition.
  */
 
+#include <egt/app.h>
 #include <egt/detail/enum.h>
 #include <egt/detail/meta.h>
 #include <egt/event.h>
@@ -19,6 +20,7 @@
 #include <egt/geometry.h>
 #include <egt/object.h>
 #include <egt/palette.h>
+#include <egt/screen.h>
 #include <egt/serialize.h>
 #include <egt/signal.h>
 #include <egt/theme.h>
@@ -214,7 +216,7 @@ public:
      * @param[in] rect The rectangle to draw.
      *
      */
-    virtual void draw(Painter& painter, const Rect& rect) = 0;
+    virtual void draw(Painter& painter, const Rect& rect);
 
     /**
      * Handle an event.
@@ -616,13 +618,19 @@ public:
     /**
      * Mark the specified rect as a damaged area of the widget.
      *
-     * @note This call will propagate to a top level parent frame that owns a
+     * @note This call will propagate to a top level parent widget that owns a
      * Screen.
      *
      * @warning damage() cannot be called while in or below a Widget::draw()
      * method.
      *
      * @param rect The rectangle to save for damage.
+     */
+    /**
+     * This will merge the damaged area with any already existing damaged area
+     * that it overlaps with into a super rectangle. Then, the whole array has
+     * to be checked again to make sure the new rectangle doesn't conflict with
+     * another existing rectangle.
      */
     virtual void damage(const Rect& rect);
 
@@ -772,14 +780,14 @@ public:
                Palette::GroupId group = Palette::GroupId::normal);
 
     /**
-     * Get a pointer to the parent Frame, or nullptr if none exists.
+     * Get a pointer to the parent Widget, or nullptr if none exists.
      */
-    Frame* parent();
+    Widget* parent();
 
     /**
-     * Get a pointer to the parent Frame, or nullptr if none exists.
+     * Get a pointer to the parent Widget, or nullptr if none exists.
      */
-    EGT_NODISCARD const Frame* parent() const;
+    EGT_NODISCARD const Widget* parent() const;
 
     /**
      * Get a pointer to the Screen instance, using using a parent as
@@ -792,7 +800,7 @@ public:
     /**
      * Align the widget.
      *
-     * This will align the widget relative to the box of its parent frame.
+     * This will align the widget relative to the box of its parent widget.
      *
      * @param[in] a The AlignFlags.
      */
@@ -1223,6 +1231,9 @@ public:
      * This will cause the widget to layout itself and any of its children.
      * This can mean, for example, the widget will resize() itself to respect
      * its min_size_hint().
+     *
+     * @note Remember that when overriding this function as a Frame, you must
+     * call layout on each child Frame to propagate the layout.
      */
     virtual void layout();
 
@@ -1388,6 +1399,191 @@ public:
 
     ~Widget() noexcept override;
 
+    /**
+     * Remove a child widget.
+     *
+     * The inverse of this call is Frame::add().
+     *
+     * @param widget The widget.
+     */
+    virtual void remove(Widget* widget);
+
+    /**
+     * Special variation of damage() that is to be called explicitly by child
+     * widgets.
+     */
+    virtual void damage_from_child(const Rect& rect)
+    {
+        damage(rect);
+    }
+
+    /**
+     * Cause the widget to draw itself and all of its children.
+     *
+     * @warning Normally this should not be called directly and instead the
+     * event loop will call this function.
+     */
+    virtual void begin_draw()
+    {
+        assert(0);
+    }
+
+    /**
+     * Convert a point with an origin of the current widget to child origin.
+     */
+    EGT_NODISCARD virtual Point to_child(const Point& p) const
+    {
+        return p - point();
+    }
+
+    /**
+     * @see Widget::to_child();
+     */
+    EGT_NODISCARD Rect to_child(Rect rect) const
+    {
+        rect.point(to_child(rect.point()));
+        return rect;
+    }
+
+    /**
+     * Convert a local point to the coordinate system of the current panel.
+     *
+     * In other words, work towards a parent that has the panel so we can get
+     * this point relative to the origin of the panel we are drawing to.
+     */
+    virtual Point to_panel(const Point& p);
+
+    /**
+     * Move the specified widget zorder down relative to other widgets with the
+     * same parent.
+     *
+     * @param widget The widget.
+     */
+    void zorder_down(const Widget* widget);
+
+    /**
+     * Move the specified widget zorder up relative to other widgets with the
+     * same parent.
+     *
+     * @param widget The widget.
+     */
+    void zorder_up(const Widget* widget);
+
+    /**
+     * Move the specified widget zorder to the bottom of the current list of widgets
+     * with the same parent.
+     *
+     * @param widget The widget.
+     */
+    void zorder_bottom(const Widget* widget);
+
+    /**
+     * Move the specified widget zorder to the top of the current list of widgets
+     * with the same parent.
+     *
+     * @param widget The widget.
+     */
+    void zorder_top(const Widget* widget);
+
+    /**
+     * Get the zorder of the widget.
+     *
+     * @param widget The widget.
+     */
+    size_t zorder(const Widget* widget) const;
+
+    /**
+     * Set the zorder of the widget.
+     *
+     * @param widget The widget.
+     * @param rank The rank in the zorder.
+     */
+    void zorder(const Widget* widget, size_t rank);
+
+    /**
+     * Add damage to the damage array.
+     *
+     * Damage rects added here must have origin at point() of this widget. This
+     * is done for several reasons, including if this widget is moved, all of the
+     * damage rects are still valid.
+     */
+    void add_damage(const Rect& rect);
+
+    /**
+     * Helper type that defines the special draw child callback.
+     */
+    using ChildDrawCallback = std::function<void(Painter& painter, Widget* widget)>;
+
+    /**
+     * Get the special child draw callback.
+     */
+    EGT_NODISCARD ChildDrawCallback special_child_draw_callback() const
+    {
+        return m_special_child_draw_callback;
+    }
+
+    /**
+     * Set the special child draw callback.
+     */
+    void special_child_draw_callback(ChildDrawCallback func)
+    {
+        m_special_child_draw_callback = std::move(func);
+    }
+
+    /**
+     * Special draw function that can be invoked when drawing each child.
+     *
+     * This is useful, for example, to draw a custom bounding box around
+     * children or to modify how a child draws by overwriting it.
+     *
+     * @param painter An instance of the Painter to use.
+     * @param widget The widget.
+     */
+    void special_child_draw(Painter& painter, Widget* widget)
+    {
+        if (m_special_child_draw_callback)
+            m_special_child_draw_callback(painter, widget);
+        else if (parent())
+            parent()->special_child_draw(painter, widget);
+    }
+
+    /**
+     * Does this Widget have a screen?
+     */
+    EGT_NODISCARD virtual bool has_screen() const { return false; }
+
+    /**
+     * Starting from this Widget, find the Widget that has a Screen.
+     *
+     * This searches up the widget hierarchy from here.
+     */
+    Widget* find_screen()
+    {
+        if (has_screen())
+            return this;
+
+        if (parent())
+            return parent()->find_screen();
+
+        return nullptr;
+    }
+
+    /**
+     * Starting from this Widget, find the Widget that has a Screen.
+     *
+     * This searches up the widget hierarchy from here.
+     */
+    EGT_NODISCARD const Widget* find_screen() const
+    {
+        if (has_screen())
+            return this;
+
+        if (parent())
+            return parent()->find_screen();
+
+        return nullptr;
+    }
+
 protected:
 
     /**
@@ -1412,7 +1608,7 @@ protected:
     /**
      * Set this widget's parent.
      */
-    virtual void set_parent(Frame* parent);
+    virtual void set_parent(Widget* parent);
 
     /**
      * Get the local box which is the same size as box(), but with the
@@ -1436,9 +1632,9 @@ protected:
     /**
      * Pointer to this widget's parent.
      *
-     * The parent is a Frame, which is capable of managing children.
+     * The parent is a Widget, which is capable of managing children.
      */
-    Frame* m_parent{nullptr};
+    Widget* m_parent{nullptr};
 
     /**
      * Unique ID of this widget.
@@ -1476,6 +1672,24 @@ protected:
      * virtual/overridden methods are executed.
      */
     void deserialize_leaf(Serializer::Properties& props);
+
+    /// @private
+    void draw_child(Painter& painter, const Rect& crect, Widget* child);
+
+    /// Used internally for calling the special child draw function.
+    ChildDrawCallback m_special_child_draw_callback;
+
+    /// Helper type for an array of children.
+    using ChildrenArray = std::vector<std::shared_ptr<Widget>>;
+
+    /// Array of child widgets in the order they were added.
+    ChildrenArray m_children;
+
+    /// The damage array for this widget.
+    Screen::DamageArray m_damage;
+
+    /// Status for whether this widget is currently drawing.
+    bool m_in_draw{false};
 
 private:
